refactor: initial separation into files
