+++ src/components/Tabs.jsx
@@ -3,7 +3,7 @@ import { ErrorBoundary } from 'react-error-boundary'
 import { Typography } from 'antd'
 
 import OptimizerTab from 'components/optimizerTab/OptimizerTab'
-import ImportTab from 'components/ImportTab'
+import ImportTab from './importTab/ImportTab'
 import RelicsTab from 'components/RelicsTab'
 import CharacterTab from 'components/CharacterTab'
 import RelicScorerTab from 'components/RelicScorerTab'
