+++ src/components/importerTab/importerTabUiConstants.ts
@@ -1,0 +1,2 @@
+export const importerTabButtonWidth = 250
+export const importerTabSpinnerMs = 500
